requirements.txt
fastapi
uvicorn
python-multipart
pillow
numpy
scikit-learn
matplotlib
seaborn
streamlit
tensorflow
opencv-python
jupyter
requests